♻️ refactor(viewmodel): split SettingsViewModel into focused sub-ViewModels
Resolves #157

SettingsViewModel was mixing 4 unrelated concerns (436 lines).
Extracted into dedicated sub-ViewModels in ViewModels/Settings/:

- AppearanceViewModel: theme and language selection
- StorageViewModel: data path management (choose/reset)
- BackupViewModel: backup, restore, browse backups, CSV export
- AboutViewModel: app version, build info, library list

SettingsViewModel is now a thin facade exposing the 4 sub-VMs as
properties (Appearance, Storage, Backup, About).

SettingsPage.xaml bindings updated to use sub-VM prefix
(e.g. Appearance.SelectedThemeIndex, Backup.LastBackupInfo).

MauiProgram.cs registers all 4 sub-VMs as transient services.

All 170 tests pass. Existing UI behaviour unchanged.

Affected: SettingsViewModel.cs, ViewModels/Settings/*.cs,
SettingsPage.xaml, MauiProgram.cs
